Fix Jinja template role alternation error for LMStudio models
Restructures message conversion to support Jinja templates with strict
role alternation requirements:

- Merge consecutive user messages to ensure user/assistant alternation
  after the optional system message

This resolves the error "conversation roles must alternate user and
assistant roles except for tool calls and results" when using models
with strict Jinja prompt templates like Mistral's devstral variants.
