+++ src/ai/utils/conversion.ts
@@ -130,91 +130,3 @@ export function convertToolResultToString(content: unknown): string {
 	return String(content);
 }
 
-// Converts VS Code LanguageModelChatRequestMessage array to AI SDK ModelMessage array
-// borrowed and adapted from https://github.com/jaykv/modelbridge/blob/main/src/provider.ts (MIT License)
-export function LM2VercelMessage(messages: readonly LanguageModelChatRequestMessage[]): ModelMessage[] {
-	logger.debug(`Converting VS Code chat messages to AI SDK format`);
-	const messagesPayload: ModelMessage[] = [];
-
-	for (const message of messages) {
-		if (message.role === LanguageModelChatMessageRole.System) {
-			logger.debug(`Processing system message`);
-			messagesPayload.push({
-				role: "system",
-				content: (message.content[0] as LanguageModelTextPart).value,
-			} as SystemModelMessage);
-		}
-		if (message.role === LanguageModelChatMessageRole.User) {
-			logger.debug(`Processing user message`);
-			const textParts: string[] = [];
-			const toolResults: ToolResultPart[] = [];
-
-			for (const part of message.content) {
-				if (part instanceof LanguageModelToolResultPart) {
-					logger.debug(`Processing tool result part with callId "${part.callId}"`);
-					toolResults.push({
-						type: "tool-result",
-						toolCallId: part.callId,
-						toolName: (part as { name?: string }).name ?? "unknown",
-						output: { type: "text", value: convertToolResultToString(part.content[0]) },
-					});
-				} else if (part instanceof LanguageModelTextPart) {
-					textParts.push(part.value);
-				}
-			}
-
-			if (toolResults.length > 0) {
-				messagesPayload.push({ role: "tool", content: toolResults } as ToolModelMessage);
-			} else {
-				messagesPayload.push({ role: "user", content: textParts.join("\n") } as UserModelMessage);
-			}
-		}
-
-		if (message.role === LanguageModelChatMessageRole.Assistant) {
-			logger.debug(`Processing assistant message`);
-			const contentParts: (TextPart | ToolCallPart | ReasoningOutput)[] = [];
-
-			for (const part of message.content) {
-				if (part instanceof LanguageModelToolCallPart) {
-					const toolCallPart: ToolCallPartWithProviderOptions = {
-						type: "tool-call",
-						toolCallId: part.callId,
-						toolName: part.name,
-						input: part.input,
-					};
-
-					// Retrieve providerMetadata from cache (e.g., Google's thoughtSignature)
-					// The metadata was stored by the provider's generateStreamingResponse
-					// Note: We do NOT delete the cache entry here because the same assistant message
-					// will be converted multiple times as part of conversation history in future turns
-					//
-					// IMPORTANT: We use providerOptions (not providerMetadata) when SENDING to providers
-					// providerMetadata is what we RECEIVE from providers, providerOptions is what we SEND
-					const cache = CacheRegistry.getCache("toolCallMetadata");
-					const cachedMetadata = cache.get(part.callId) as ToolCallMetadata | undefined;
-					if (cachedMetadata?.providerMetadata) {
-						logger.debug(`Using cached provider metadata for toolCallId "${part.callId}"`);
-						toolCallPart.providerOptions = cachedMetadata.providerMetadata;
-					}
-
-					contentParts.push(toolCallPart);
-				} else if (part instanceof LanguageModelThinkingPart) {
-					if (part.id && part.id.startsWith("error")) {
-						continue;
-						// Specialized thinking part for error messages; skip and dont include in context.
-						// VScode doesn't allow a reasonable way to inject errors into the chat history.
-						// So in generateStreamingResponse() we use a thinking part with id "error" to indicate an error.
-						// And specifically exclude it here.
-					}
-					const text = Array.isArray(part.value) ? part.value.join("") : part.value;
-					contentParts.push({ type: "reasoning", text });
-				} else if (part instanceof LanguageModelTextPart) {
-					contentParts.push({ type: "text", text: part.value });
-				}
-			}
-
-			messagesPayload.push({ role: "assistant", content: contentParts } as AssistantModelMessage);
-		}
-	}
-	return messagesPayload;
-}
